Finished makefile
Finished the makefile. Changed it so that new files are created inside
of a new directory and running clean_all cleans out this directory.
(Requires the user to make a directory called ‘tests’)
Also changed one line of code in ben’s source code so that it would
properly compile. Still looking into that issue though