more cleanup and get_val_after_reassign
did more cleanup and tried out a couple tests that failed to produce
any bugs. however, get_val_after_reassign (which is the counterpart to
the existing get_size_after_reassign) exposes a bug in the val pointer
returned by cache_get. it looks like some caches are reusing the same
pointer so that if we reassign a value, the first output gets
reassigned as well because all outs share the same address.