another LRU test
tests for effects of trying to update a function with a value that
exceeds maxmem by itself. as of right now i want it to not evict
anything and also not add the value to the cache. i’m seeing evidence
of both problems in certain caches, though there might be some
ambiguity regarding whether or not we evict the respective key so this
is up in the air right now