adding/editing tests
purpose of this branch is mainly to update existing tests so that they
make use of our testing kit. but i am also adding new tests as i go to
make sure the bugs that each test raises are as independent as
possible. will make the writeup a bit longer but it will probably be
clearer as to what behavior is causing what bugs (i.e. i want to limit
as many duplicated bugs/know the sources of bugs so that i know the
dependencies in all our tests. should help minimize false positives and
false negatives, which i’ve already been finding.)